Move pages into src to enable posts routes
